Extract LevelIcon module
